Add files via upload
My Undergraduate Thesis that managed to bag the Gold Thesis Award. 
This uses the Arduino Uno with the MPU 6050 Accelerometer, where the sensor would get the inclination of its position at your back. If you stay in bad posture for too long (adjustable), it either sounds an alarm or vibrates (can be set).
It also interfaces with a PC or Android Exercise Game, and there's a database in both to show you which days you exercised. As of uploading, I still don't know how to upload these 2. Github says that either there's too many files, or it's too big.